src/core/compliance.py
from typing import Dict, Any, List
import logging
import yaml
import os
from datetime import datetime

logger = logging.getLogger(__name__)

class ComplianceValidator:
    def __init__(self):
        self.standards = {
            'owasp': self._load_owasp_standards(),
            'pci_dss': self._load_pci_dss_standards(),
            'gdpr': self._load_gdpr_standards()
        }
        self.compliance_status = {}

    def _load_owasp_standards(self) -> Dict[str, Any]:
        """Load OWASP security standards"""
        return {
            'A1': {
                'name': 'Injection',
                'requirements': [
                    'Validate all user inputs',
                    'Use parameterized queries',
                    'Implement input sanitization'
                ]
            },
            'A2': {
                'name': 'Broken Authentication',
                'requirements': [
                    'Implement strong password policies',
                    'Use multi-factor authentication',
                    'Secure session management'
                ]
            },
            'A3': {
                'name': 'Sensitive Data Exposure',
                'requirements': [
                    'Encrypt sensitive data',
                    'Use secure protocols',
                    'Implement proper key management'
                ]
            }
        }

    def _load_pci_dss_standards(self) -> Dict[str, Any]:
        """Load PCI DSS standards"""
        return {
            '1': {
                'name': 'Install and maintain a firewall',
                'requirements': [
                    'Establish firewall configuration',
                    'Prohibit direct public access',
                    'Implement DMZ'
                ]
            },
            '2': {
                'name': 'Do not use vendor defaults',
                'requirements': [
                    'Change default passwords',
                    'Remove unnecessary accounts',
                    'Disable unnecessary services'
                ]
            }
        }

    def _load_gdpr_standards(self) -> Dict[str, Any]:
        """Load GDPR standards"""
        return {
            '1': {
                'name': 'Data Protection',
                'requirements': [
                    'Implement data encryption',
                    'Maintain data processing records',
                    'Conduct privacy impact assessments'
                ]
            },
            '2': {
                'name': 'User Rights',
                'requirements': [
                    'Implement right to access',
                    'Implement right to erasure',
                    'Implement data portability'
                ]
            }
        }

    def check_compliance(self) -> Dict[str, Any]:
        """Check compliance with all standards"""
        try:
            results = {}
            
            for standard, requirements in self.standards.items():
                results[standard] = self._check_standard_compliance(standard, requirements)
            
            self.compliance_status = results
            return results
        except Exception as e:
            logger.error(f"Error checking compliance: {str(e)}")
            return {}

    def _check_standard_compliance(self, standard: str, requirements: Dict[str, Any]) -> Dict[str, Any]:
        """Check compliance with a specific standard"""
        try:
            results = {
                'status': 'compliant',
                'last_check': datetime.utcnow().isoformat(),
                'requirements': {}
            }
            
            for req_id, req_data in requirements.items():
                req_status = self._check_requirement(req_data['requirements'])
                results['requirements'][req_id] = {
                    'name': req_data['name'],
                    'status': req_status,
                    'details': req_data['requirements']
                }
                
                if not req_status:
                    results['status'] = 'non-compliant'
            
            return results
        except Exception as e:
            logger.error(f"Error checking {standard} compliance: {str(e)}")
            return {'status': 'error', 'error': str(e)}

    def _check_requirement(self, requirements: List[str]) -> bool:
        """Check if a specific requirement is met"""
        # This is a placeholder implementation
        # In a real implementation, this would check actual system configuration
        return True

    def generate_compliance_report(self) -> Dict[str, Any]:
        """Generate a detailed compliance report"""
        try:
            if not self.compliance_status:
                self.check_compliance()
            
            report = {
                'timestamp': datetime.utcnow().isoformat(),
                'overall_status': 'compliant',
                'standards': {}
            }
            
            for standard, results in self.compliance_status.items():
                report['standards'][standard] = results
                if results['status'] == 'non-compliant':
                    report['overall_status'] = 'non-compliant'
            
            return report
        except Exception as e:
            logger.error(f"Error generating compliance report: {str(e)}")
            return {}

    def export_compliance_report(self, format: str = 'yaml') -> str:
        """Export compliance report in specified format"""
        try:
            report = self.generate_compliance_report()
            
            if format.lower() == 'yaml':
                return yaml.dump(report, default_flow_style=False)
            elif format.lower() == 'json':
                import json
                return json.dumps(report, indent=2)
            else:
                raise ValueError(f"Unsupported format: {format}")
        except Exception as e:
            logger.error(f"Error exporting compliance report: {str(e)}")
            return "" 