requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
requests==2.26.0
PyYAML==5.4.1
elasticsearch==7.14.0
prometheus-client==0.11.0
gunicorn==20.1.0
pytest==6.2.5
black==21.9b0
flake8==3.9.2
mypy==0.910
pydantic==1.8.2
redis==3.5.3
celery==5.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0 